move client validation
