--- client/prompt_validation_test.go
@@ -1,16 +1,0 @@
-package client_test
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/require"
-
-	"github.com/cosmos/cosmos-sdk/client"
-)
-
-func TestValidatePromptURL(t *testing.T) {
-	require := require.New(t)
-
-	require.NoError(client.ValidatePromptURL("https://example.com"))
-	require.ErrorContains(client.ValidatePromptURL("foo"), "invalid URL")
-}
